Using qt types.

+++ source/host/service_main.cc
@@ -174,7 +174,7 @@ std::optional<QString> currentSessionName()
 
     // We are looking for the current session in the sorted list.
     size_t user_number = 0;
-    for (size_t i = 0; i < times.size(); ++i)
+    for (int i = 0; i < times.size(); ++i)
     {
         if (times[i].first == current_session_info.sessionId())
         {
